SocialX Backend/requirements.txt
Django
djangorestframework
requests
python-decouple
